+++ backend/src/main/java/ma/dnaengineering/backend/BackendApplication.java
@@ -1,20 +1,15 @@
 package ma.dnaengineering.backend;
 
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class BackendApplication implements CommandLineRunner {
+public class BackendApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(BackendApplication.class, args);
 	}
 
-	@Override
-	public void run(String... args) throws Exception {
 	
-
-	}
 
 }
