+++ backend/src/main/java/ma/dnaengineering/backend/dao/EmployeeDao.java
@@ -4,10 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import ma.dnaengineering.backend.model.Employee;
+
 @Repository
 public interface EmployeeDao extends JpaRepository<Employee,Long> {
-
-   
-
+    
 }
