+++ backend/src/main/java/ma/dnaengineering/backend/model/Employee.java
@@ -1,8 +1,6 @@
 package ma.dnaengineering.backend.model;
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -14,10 +12,7 @@ import lombok.NoArgsConstructor;
 @Entity
 public class Employee {
 
-   // id,employee_name,job_title,salary
-
    @Id
-   @GeneratedValue(strategy = GenerationType.AUTO)
    private Long id;
    private String employeeName; 
    private String jobTitle;
